refactor(test): the fake workbench tab is four modules and the page's own run helper
`openFakeWorkbenchTab` was one 485 line function holding five concerns, and the
file was 615 lines. Both are past the limits this repo sets for new code, and
the review named the split. The concerns are now their own modules, each taking
what it needs as an argument:

- `fake-tab-document.ts`: the load and the save, with the page's refusal rules.
- `fake-tab-run.ts`: the execute request and the stream drain.
- `fake-tab-handlers.ts`: the handler table, built from the manifest.
- `fake-tab-ui-actions.ts`: the browser leg, from claim to completion record.

`openFakeWorkbenchTab` is the wiring that holds them together. Every file is
under 300 lines and the bodies did not change.

The run-id answer was a second copy of the page's own, budget included, and the
page's copy left a rejection unhandled: `finally` passes the rejection on and
`void` drops it. One `startAndIdentifyRun` now serves both, the tab imports it
the way it imports the rest of the page, and a rejected start answers the caller
instead of escaping.

`seedComparisonWorkbench` was 93 source lines; the two carrier shapes and the
saved judge are module-level builders now, the way the file already builds its
baseline target.

Also: the boolean on a turn's text carries the prefix the repo asks for, an
empty label is not a verdict in `expectComparisonScored` any more (it already
was not in `expectRunHasRealScores`), and the suite timeout states the
arithmetic it claims and covers it.

Verified against the live stack: `workbench-fake-tab.harness.test.ts` 2/2,
both carrier shapes, no agent turns.
