+++ platform/app/src/pages/[project]/experiments/workbench/[slug].tsx
@@ -20,6 +20,7 @@ import { TableSettingsMenu } from "~/experiments-v3/components/TableSettingsMenu
 import { UndoRedo } from "~/experiments-v3/components/UndoRedo";
 import { VersionHistoryButton } from "~/experiments-v3/components/VersionHistoryButton";
 import { WorkbenchStaleBanner } from "~/experiments-v3/components/WorkbenchStaleBanner";
+import { startAndIdentifyRun } from "~/experiments-v3/execution/runIdentification";
 import { useAutosaveEvaluationsV3 } from "~/experiments-v3/hooks/useAutosaveEvaluationsV3";
 import { useEvaluationsV3Store } from "~/experiments-v3/hooks/useEvaluationsV3Store";
 import { useExecuteEvaluation } from "~/experiments-v3/hooks/useExecuteEvaluation";
@@ -49,39 +50,6 @@ import { useOrganizationTeamProject } from "~/hooks/useOrganizationTeamProject";
 import { api } from "~/utils/api";
 import { useRouter } from "~/utils/compat/next-router";
 import { assertCrispChatHidden } from "~/utils/crispBubblePolicy";
-
-/**
- * How long the run action waits for the stream to name its run.
- *
- * The id is minted server-side and travels on the first frame, so this only
- * has to cover opening the connection — the same budget `fetchSSE` gives that
- * connection. The run itself takes minutes and is never waited for.
- */
-const RUN_ID_WAIT_MS = 30_000;
-
-/**
- * Start a run and answer with its id.
- *
- * Answers with no id rather than holding the action open when the stream ends,
- * fails, or never opens: a run the caller cannot name is still a run that is
- * going, and the action's own budget is not the place to discover otherwise.
- */
-function startAndIdentifyRun(
-  start: (onRunStarted: (runId: string) => void) => Promise<void>,
-): Promise<string | undefined> {
-  return new Promise((resolve) => {
-    let settled = false;
-    let timer: ReturnType<typeof setTimeout> | undefined;
-    const answer = (runId?: string) => {
-      if (settled) return;
-      settled = true;
-      if (timer) clearTimeout(timer);
-      resolve(runId);
-    };
-    timer = setTimeout(() => answer(undefined), RUN_ID_WAIT_MS);
-    void Promise.resolve(start(answer)).finally(() => answer(undefined));
-  });
-}
 
 /**
  * Experiments Workbench Page
@@ -514,9 +482,10 @@ export default function ExperimentsWorkbenchPage() {
           //
           // The payload-to-scope mapping is shared with that fallback so both
           // cover the same cells.
-          const runId = await startAndIdentifyRun((onRunStarted) =>
-            executeEvaluation(scopeFromRunPayload(payload), { onRunStarted }),
-          );
+          const runId = await startAndIdentifyRun({
+            start: (onRunStarted) =>
+              executeEvaluation(scopeFromRunPayload(payload), { onRunStarted }),
+          });
           return { runId, status: "running" as const };
         } finally {
           setActionActivity(null);
